+++ AffindaAPI/AffindaAPI/Models/FieldDeprecated.Serialization.cs
@@ -11,13 +11,18 @@ using Azure.Core;
 
 namespace Affinda.API.Models
 {
-    public partial class Field : IUtf8JsonSerializable
+    public partial class FieldDeprecated : IUtf8JsonSerializable
     {
         void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
         {
             writer.WriteStartObject();
             writer.WritePropertyName("label");
             writer.WriteStringValue(Label);
+            if (Optional.IsDefined(Slug))
+            {
+                writer.WritePropertyName("slug");
+                writer.WriteStringValue(Slug);
+            }
             writer.WritePropertyName("dataPoint");
             writer.WriteStringValue(DataPoint);
             if (Optional.IsDefined(Mandatory))
@@ -25,17 +30,15 @@ namespace Affinda.API.Models
                 writer.WritePropertyName("mandatory");
                 writer.WriteBooleanValue(Mandatory.Value);
             }
+            if (Optional.IsDefined(Disabled))
+            {
+                writer.WritePropertyName("disabled");
+                writer.WriteBooleanValue(Disabled.Value);
+            }
             if (Optional.IsDefined(AutoValidationThreshold))
             {
-                if (AutoValidationThreshold != null)
-                {
-                    writer.WritePropertyName("autoValidationThreshold");
-                    writer.WriteNumberValue(AutoValidationThreshold.Value);
-                }
-                else
-                {
-                    writer.WriteNull("autoValidationThreshold");
-                }
+                writer.WritePropertyName("autoValidationThreshold");
+                writer.WriteNumberValue(AutoValidationThreshold.Value);
             }
             if (Optional.IsDefined(ShowDropdown))
             {
@@ -55,19 +58,26 @@ namespace Affinda.API.Models
             writer.WriteEndObject();
         }
 
-        internal static Field DeserializeField(JsonElement element)
+        internal static FieldDeprecated DeserializeFieldDeprecated(JsonElement element)
         {
             string label = default;
+            Optional<string> slug = default;
             string dataPoint = default;
             Optional<bool> mandatory = default;
-            Optional<float?> autoValidationThreshold = default;
+            Optional<bool> disabled = default;
+            Optional<float> autoValidationThreshold = default;
             Optional<bool> showDropdown = default;
-            Optional<IList<Field>> fields = default;
+            Optional<IList<FieldDeprecated>> fields = default;
             foreach (var property in element.EnumerateObject())
             {
                 if (property.NameEquals("label"))
                 {
                     label = property.Value.GetString();
+                    continue;
+                }
+                if (property.NameEquals("slug"))
+                {
+                    slug = property.Value.GetString();
                     continue;
                 }
                 if (property.NameEquals("dataPoint"))
@@ -85,11 +95,21 @@ namespace Affinda.API.Models
                     mandatory = property.Value.GetBoolean();
                     continue;
                 }
+                if (property.NameEquals("disabled"))
+                {
+                    if (property.Value.ValueKind == JsonValueKind.Null)
+                    {
+                        property.ThrowNonNullablePropertyIsNull();
+                        continue;
+                    }
+                    disabled = property.Value.GetBoolean();
+                    continue;
+                }
                 if (property.NameEquals("autoValidationThreshold"))
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
-                        autoValidationThreshold = null;
+                        property.ThrowNonNullablePropertyIsNull();
                         continue;
                     }
                     autoValidationThreshold = property.Value.GetSingle();
@@ -112,16 +132,16 @@ namespace Affinda.API.Models
                         property.ThrowNonNullablePropertyIsNull();
                         continue;
                     }
-                    List<Field> array = new List<Field>();
+                    List<FieldDeprecated> array = new List<FieldDeprecated>();
                     foreach (var item in property.Value.EnumerateArray())
                     {
-                        array.Add(DeserializeField(item));
+                        array.Add(DeserializeFieldDeprecated(item));
                     }
                     fields = array;
                     continue;
                 }
             }
-            return new Field(label, dataPoint, Optional.ToNullable(mandatory), Optional.ToNullable(autoValidationThreshold), Optional.ToNullable(showDropdown), Optional.ToList(fields));
+            return new FieldDeprecated(label, slug.Value, dataPoint, Optional.ToNullable(mandatory), Optional.ToNullable(disabled), Optional.ToNullable(autoValidationThreshold), Optional.ToNullable(showDropdown), Optional.ToList(fields));
         }
     }
 }