+++ AffindaAPI/AffindaAPI/Models/FieldDeprecated.cs
@@ -11,14 +11,14 @@ using Azure.Core;
 
 namespace Affinda.API.Models
 {
-    /// <summary> The Field. </summary>
-    public partial class Field
+    /// <summary> The FieldDeprecated. </summary>
+    public partial class FieldDeprecated
     {
-        /// <summary> Initializes a new instance of Field. </summary>
+        /// <summary> Initializes a new instance of FieldDeprecated. </summary>
         /// <param name="label"></param>
-        /// <param name="dataPoint"> Data point identifier. </param>
+        /// <param name="dataPoint"></param>
         /// <exception cref="ArgumentNullException"> <paramref name="label"/> or <paramref name="dataPoint"/> is null. </exception>
-        public Field(string label, string dataPoint)
+        public FieldDeprecated(string label, string dataPoint)
         {
             if (label == null)
             {
@@ -31,21 +31,25 @@ namespace Affinda.API.Models
 
             Label = label;
             DataPoint = dataPoint;
-            Fields = new ChangeTrackingList<Field>();
+            Fields = new ChangeTrackingList<FieldDeprecated>();
         }
 
-        /// <summary> Initializes a new instance of Field. </summary>
+        /// <summary> Initializes a new instance of FieldDeprecated. </summary>
         /// <param name="label"></param>
-        /// <param name="dataPoint"> Data point identifier. </param>
+        /// <param name="slug"></param>
+        /// <param name="dataPoint"></param>
         /// <param name="mandatory"></param>
+        /// <param name="disabled"></param>
         /// <param name="autoValidationThreshold"></param>
         /// <param name="showDropdown"></param>
         /// <param name="fields"></param>
-        internal Field(string label, string dataPoint, bool? mandatory, float? autoValidationThreshold, bool? showDropdown, IList<Field> fields)
+        internal FieldDeprecated(string label, string slug, string dataPoint, bool? mandatory, bool? disabled, float? autoValidationThreshold, bool? showDropdown, IList<FieldDeprecated> fields)
         {
             Label = label;
+            Slug = slug;
             DataPoint = dataPoint;
             Mandatory = mandatory;
+            Disabled = disabled;
             AutoValidationThreshold = autoValidationThreshold;
             ShowDropdown = showDropdown;
             Fields = fields;
@@ -53,15 +57,19 @@ namespace Affinda.API.Models
 
         /// <summary> Gets or sets the label. </summary>
         public string Label { get; set; }
-        /// <summary> Data point identifier. </summary>
+        /// <summary> Gets or sets the slug. </summary>
+        public string Slug { get; set; }
+        /// <summary> Gets or sets the data point. </summary>
         public string DataPoint { get; set; }
         /// <summary> Gets or sets the mandatory. </summary>
         public bool? Mandatory { get; set; }
+        /// <summary> Gets or sets the disabled. </summary>
+        public bool? Disabled { get; set; }
         /// <summary> Gets or sets the auto validation threshold. </summary>
         public float? AutoValidationThreshold { get; set; }
         /// <summary> Gets or sets the show dropdown. </summary>
         public bool? ShowDropdown { get; set; }
         /// <summary> Gets the fields. </summary>
-        public IList<Field> Fields { get; }
+        public IList<FieldDeprecated> Fields { get; }
     }
 }